agent/requirements.txt
google-genai==0.8.3
websockets==13.1
python-dotenv==1.0.1
